Create grid, reset button and raw functionality
HTML only includes buttons and the grid container
Grid is created within the javascript when the DOM is loaded.
Every row is an element and every row contains the set amount of elements
defined by the size via display flex.

It draws when you hover over the pixels, while the mouse is clicked.
This is created by an event listener that listens to mousedown and
mouseup events on the whole body.

Resetting the grid works via iterating over through the PIXELS array